seafly/requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
Django==3.0.4
django-debug-toolbar==2.2
form==0.0.1
gunicorn==20.0.4
markdown2==2.3.8
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
